# bins/delonix-runtime-bin/templates/nestjs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/health/live": {
      "get": {
        "operationId": "HealthController_live",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Alive.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusBody"
                }
              }
            }
          }
        },
        "summary": "Liveness — the process answers.",
        "tags": [
          "health"
        ]
      }
    },
    "/api/v1/health/ready": {
      "get": {
        "operationId": "HealthController_ready",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Ready.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusBody"
                }
              }
            }
          },
          "503": {
            "description": "Starting, draining, or a required dependency is failing.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusBody"
                }
              }
            }
          }
        },
        "summary": "Readiness — the process should receive traffic.",
        "tags": [
          "health"
        ]
      }
    },
    "/api/v1/notes": {
      "post": {
        "operationId": "NotesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created. The Location header points at the note.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "413": {
            "description": "Body over HTTP_MAX_BODY_BYTES.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "summary": "Create a note.",
        "tags": [
          "notes"
        ]
      },
      "get": {
        "operationId": "NotesController_list",
        "parameters": [
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One page.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotePageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid offset or limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "summary": "List notes, newest first.",
        "tags": [
          "notes"
        ]
      }
    },
    "/api/v1/notes/{id}": {
      "get": {
        "operationId": "NotesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The note.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteResponse"
                }
              }
            }
          },
          "404": {
            "description": "No note with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "summary": "Get one note.",
        "tags": [
          "notes"
        ]
      }
    },
    "/api/v1/webhooks/inbound": {
      "post": {
        "operationId": "InboundWebhookController_receive",
        "parameters": [
          {
            "name": "webhook-signature",
            "in": "header",
            "description": "v1,<base64 HMAC-SHA256>",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "webhook-timestamp",
            "in": "header",
            "description": "Unix seconds.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "webhook-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookEventDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Accepted, or a duplicate of an accepted delivery."
          },
          "400": {
            "description": "Missing headers or a body that is not JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Signature or timestamp not valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "413": {
            "description": "Body over 256 KiB.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Unknown event, or a note.create the use case refused.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "summary": "Receive a signed webhook (Standard Webhooks). Only served when WEBHOOK_INBOUND_SECRET is set.",
        "tags": [
          "webhooks"
        ]
      }
    }
  },
  "info": {
    "title": "__NAME__",
    "description": "HTTP contract of __NAME__. Every error answers {\"error\":{\"code\",\"message\",\"fields\"?,\"request_id\"}}. Breaking changes go to a new /api/v2 prefix.",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "StatusBody": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "ready"
          },
          "dependency": {
            "type": "string",
            "description": "The failing dependency, when there is one."
          }
        },
        "required": [
          "status"
        ]
      },
      "CreateNoteDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "Trimmed; 1–200 characters."
          },
          "body": {
            "type": "string",
            "maxLength": 10000
          }
        },
        "required": [
          "title"
        ]
      },
      "NoteResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "title",
          "body",
          "created_at"
        ]
      },
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "validation_failed",
            "description": "validation_failed (422), malformed_json (400), body_too_large (413), not_found (404), internal (500), and the webhook codes missing_signature (400), invalid_signature (401), unknown_event (422)."
          },
          "message": {
            "type": "string",
            "example": "invalid input"
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "example": {
              "title": "is required"
            }
          },
          "request_id": {
            "type": "string",
            "example": "4f1c2d3e4b5a6978"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ]
      },
      "ErrorBody": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ]
      },
      "NotePageResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NoteResponse"
            }
          },
          "total": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ]
      },
      "WebhookEventDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "ping",
              "note.create"
            ]
          },
          "data": {
            "description": "Required for note.create.",
            "allOf": [
              {
                "$ref": "#/components/schemas/CreateNoteDto"
              }
            ]
          }
        },
        "required": [
          "type"
        ]
      }
    }
  }
}